Add check user for delete, update, change language comments

+++ src/controllers/lists/index.js
@@ -2,9 +2,9 @@ const express = require('express');
 const db = require('../../db');
 
 const router = express.Router();
-//прочитать все листы
+// take all list
 router.get('/', (request, response) => {
-  db.all('SELECT * FROM list', [], (error, rows) => {
+  db.all('SELECT * FROM list WHERE user_id = ?', [1], (error, rows) => {
     if (error) {
       console.error(error.message);
       response.sendStatus(500);
@@ -13,9 +13,9 @@ router.get('/', (request, response) => {
     response.send(rows);
   });
 });
-//запрос листа по id
+// request list by id
 router.get('/:id', (request, response) => {
-  db.get('SELECT * FROM list WHERE id = ?', [request.params.id], (error, row) => {
+  db.get('SELECT * FROM list WHERE id = ? AND user_id = ?', [request.params.id, 1], (error, row) => {
     if (error) {
       console.error(error.message);
       response.sendStatus(404);
@@ -24,36 +24,32 @@ router.get('/:id', (request, response) => {
     response.send(row);
   });
 });
-//создание нового листа
-router.post('/create', (request, response) => {
+// create new list
+router.post('/', (request, response) => {
   db.run('INSERT INTO list(name, user_id) VALUES (?, ?)', [request.body.name, request.body.user_id], (error) => {
     if (error) {
       console.error(error.message);
       response.sendStatus(501);
-      return;
     }
   });
   response.send('SUCCESS');
 });
-//изменение по id
+// change list by id
 router.put('/:id', (request, response) => {
-  console.log(request.params.id);
-  db.run('UPDATE list SET name = ? WHERE id = ?', [request.body.name, request.params.id], (error) => {
+  db.run('UPDATE list SET name = ? WHERE id = ? AND user_id = ?', [request.body.name, request.params.id, 1], (error) => {
     if (error) {
       console.error(error.message);
       response.sendStatus(501);
-      return;
     }
   });
   response.send('List update!');
 });
-//удаление листа по id
+// delete list by id
 router.delete('/:id', (request, response) => {
-  db.run('DELETE FROM list WHERE id = ?', [request.params.id], (error) => {
+  db.run('DELETE FROM list WHERE id = ? AND user_id = ?', [request.params.id, 1], (error) => {
     if (error) {
       console.error(error.message);
       response.sendStatus(501);
-      return;
     }
   });
   response.sendStatus(410);
